fixed name bug
It turned out that I was not updating the builder if 'update builder' was not beeing clicked
which is why the name never changed. This also means that I probably made some mistakes since
all the other fields were not updating its just that I didn't notice it with the other fields.