added origin
Every record now has a origin string that determines where that payment came from
but not which profile exactly it came from. Multiple profiles can have the same
origin which allows for different inputs from the same source.

This is a step towards being able to connect records as the same transaction (#35)
as well as checking for duplicate records (#27) from the same origin